example-5-cdk-with-environments/pyproject.toml
[tool.poetry]
name = "example-5"
version = "0.1.0"
description = "Example demoing using CDK with GitHub Actions Environments"
authors = ["Ciaran Evans <ciaran@developmentseed.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"aws-cdk.core" = "^1.106.0"
boto3 = "^1.17.80"
"aws-cdk.aws-s3" = "^1.106.0"

[tool.poetry.dev-dependencies]
black = "^21.5b1"
isort = "^5.8.0"
flake8 = "^3.9.2"
pytest = "^6.2.4"
assertpy = "^1.1"
moto = "^2.0.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
